app/server/requirements.txt
flask
flask-cors
flask-session
python-dotenv
werkzeug
psycopg2-binary
celery
redis
pymupdf
pytesseract
Pillow
sentence-transformers
chromadb
duckduckgo-search
ollama
google-generativeai
gunicorn
requests